onepost/requirements/production.txt
# PRECAUTION: avoid production dependencies that aren't in development

-r base.txt

uwsgi
psycopg2-binary
sentry-sdk==2.9.0
# Django
# ------------------------------------------------------------------------------
django-storages[s3]==1.14.4
firebase_admin
django-filter

# Security
# ------------------------------------------------------------------------------
python-jose
pyotp
